Implement the EditContactDescriptor as an individual class rather than a nested static class
