+++ src/main/java/seedu/address/logic/commands/EditContactCommand.java
@@ -7,11 +7,9 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TELEGRAM;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_CONTACTS;
 
 import java.util.List;
-import java.util.Optional;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
@@ -106,80 +104,6 @@ public class EditContactCommand extends Command {
         EditContactCommand e = (EditContactCommand) other;
         return index.equals(e.index)
                 && editContactDescriptor.equals(e.editContactDescriptor);
-    }
-
-    /**
-     * Stores the details to edit the contact with. Each non-empty field value will replace the
-     * corresponding field value of the contact.
-     */
-    public static class EditContactDescriptor {
-        private Name name;
-        private Email email;
-        private Telegram telegram;
-
-        public EditContactDescriptor() {}
-
-        /**
-         * Copy constructor.
-         * A defensive copy of {@code tags} is used internally.
-         */
-        public EditContactDescriptor(EditContactDescriptor toCopy) {
-            setName(toCopy.name);
-            setEmail(toCopy.email);
-            setTelegram(toCopy.telegram);
-        }
-
-        /**
-         * Returns true if at least one field is edited.
-         */
-        public boolean isAnyFieldEdited() {
-            return CollectionUtil.isAnyNonNull(name, email, telegram);
-        }
-
-        public void setName(Name name) {
-            this.name = name;
-        }
-
-        public Optional<Name> getName() {
-            return Optional.ofNullable(name);
-        }
-
-        public void setEmail(Email email) {
-            this.email = email;
-        }
-
-        public Optional<Email> getEmail() {
-            return Optional.ofNullable(email);
-        }
-
-        public void setTelegram(Telegram telegram) {
-            this.telegram = telegram;
-        }
-
-        public Optional<Telegram> getTelegram() {
-            return Optional.ofNullable(telegram);
-        }
-
-
-        @Override
-        public boolean equals(Object other) {
-            // short circuit if same object
-            if (other == this) {
-                return true;
-            }
-
-            // instanceof handles nulls
-            if (!(other instanceof EditContactDescriptor)) {
-                return false;
-            }
-
-            // state check
-            EditContactDescriptor e = (EditContactDescriptor) other;
-
-            return getName().equals(e.getName())
-                    && getEmail().equals(e.getEmail())
-                    && getTelegram().equals(e.getTelegram());
-        }
     }
 
     @Override
